refactor: add component test
